68308 - Refacto move template service
